changed all rabbitmqserver instances
- changed the rabbitmq user
- gave each instance a development and production
