+++ VMs/database/Authentication Reciever/RabbitMQ.ini
@@ -1,11 +1,19 @@
-[server]
+[development]
 BROKER_HOST = 127.0.0.1
 BROKER_PORT = 5672
-USER = auth
+USER = bookQuest
 PASSWORD = 8bkJ3r4dWSU1lkL6HQT7
-VHOST = BookQuest
-EXCHANGE = credentialsExchange
-QUEUE = credentialsQueue
-;EXCHANGE_TYPE = ????
+VHOST = bookQuest
+EXCHANGE = authenticationExchange
+QUEUE = authenticationQueue
 AUTO_DELETE = true
 
+[production]
+BROKER_HOST = rabbitmq.tortoise-daggertooth.ts.net
+BROKER_PORT = 5672
+USER = bookQuest
+PASSWORD = 8bkJ3r4dWSU1lkL6HQT7
+VHOST = bookQuest
+EXCHANGE = authenticationExchange
+QUEUE = authenticationQueue
+AUTO_DELETE = true
