+++ VMs/dmz/README.md
@@ -1,5 +1,5 @@
 # DMZ
 
-* **`/RabbitMQServer/`** ➜ folder containing everything necessary for the DMZ to handle rabbitMQ messages
+* **`/Search Reciever/`** ➜ folder containing everything necessary for the DMZ to handle rabbitMQ messages pertaining to the search function.
   * *`rabbitMQ.ini`* ➜ configuration file with variables for the RabbitMQ connection
   * *`rabbitMQServer.php`* ➜ Currently does nothing but planned to fetch from the google api
