--- VMs/dmz/Search Broker Reciever/RabbitMQ.ini
@@ -1,11 +1,0 @@
-[server]
-BROKER_HOST = 127.0.0.1
-BROKER_PORT = 5672
-USER = auth
-PASSWORD = 8bkJ3r4dWSU1lkL6HQT7
-VHOST = BookQuest
-EXCHANGE = searchExchange
-QUEUE = searchQueue
-;EXCHANGE_TYPE = ????
-AUTO_DELETE = true
-
