+++ VMs/dmz/Search Reciever/RabbitMQ.ini
@@ -4,8 +4,8 @@ BROKER_PORT = 5672
 USER = bookQuest
 PASSWORD = 8bkJ3r4dWSU1lkL6HQT7
 VHOST = bookQuest
-EXCHANGE = authenticationExchange
-QUEUE = authenticationQueue
+EXCHANGE = searchExchange
+QUEUE = searchDmzQueue
 AUTO_DELETE = true
 
 [production]
@@ -14,6 +14,6 @@ BROKER_PORT = 5672
 USER = bookQuest
 PASSWORD = 8bkJ3r4dWSU1lkL6HQT7
 VHOST = bookQuest
-EXCHANGE = authenticationExchange
-QUEUE = authenticationQueue
+EXCHANGE = searchExchange
+QUEUE = searchDmzQueue
 AUTO_DELETE = true
