+++ VMs/dmz/Search Reciever/get_host_info.inc
