+++ VMs/dmz/Search Reciever/rabbitMQLib.inc
