+++ VMs/frontend/website/common/utils/searchDatabase/RabbitMQ.ini
@@ -1,10 +1,19 @@
-[server]
+[development]
 BROKER_HOST = 127.0.0.1
 BROKER_PORT = 5672
-USER = auth
+USER = bookQuest
 PASSWORD = 8bkJ3r4dWSU1lkL6HQT7
-VHOST = BookQuest
-EXCHANGE = databaseSearchExchange
-QUEUE = databaseSearchQueue
+VHOST = bookQuest
+EXCHANGE = searchExchange
+QUEUE = searchDatabaseQueue
 AUTO_DELETE = true
 
+[production]
+BROKER_HOST = rabbitmq.tortoise-daggertooth.ts.net
+BROKER_PORT = 5672
+USER = bookQuest
+PASSWORD = 8bkJ3r4dWSU1lkL6HQT7
+VHOST = bookQuest
+EXCHANGE = searchExchange
+QUEUE = searchDatabaseQueue
+AUTO_DELETE = true
